% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{aboutServer}
\alias{aboutServer}
\alias{aboutOutput}
\alias{aboutApp}
\title{About Server}
\usage{
aboutServer(id, helppath = NULL)

aboutOutput(id)

aboutApp(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{helppath}{path to help markdown}
}
\value{
reactive server
}
\description{
About Server
}
