% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bare.R
\name{bareServer}
\alias{bareServer}
\alias{bareInput}
\alias{bareOutput}
\alias{bareApp}
\title{Shiny bare Server for foundr Package}
\usage{
bareServer(
  id,
  traitData = NULL,
  traitSignal = NULL,
  traitStats = NULL,
  customSettings = NULL,
  traitModule = NULL
)

bareInput(id)

bareOutput(id)

bareApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}

\item{title}{title of app}
}
\value{
reactive server
}
\description{
Shiny bare Server for foundr Package
}
