% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplotApp.R
\name{biplotApp}
\alias{biplotApp}
\alias{biplotServer}
\alias{biplotOutput}
\title{BiPlot App}
\usage{
biplotApp()

biplotServer(id, panel_par, plot_par, plot_info, contrast_table)

biplotOutput(id)
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{plot_info}{reactive values from contrastPlot}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object

nothing returned
}
\description{
BiPlot App

Shiny Module Output for Contrast Plots
}
