% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{contrast_signal}
\alias{contrast_signal}
\title{Turn `conditionContrasts` object into a `traitSignal` object.}
\usage{
contrast_signal(contrasts)
}
\arguments{
\item{contrasts}{data frame}
}
\value{
data frame
}
\description{
Turn `conditionContrasts` object into a `traitSignal` object.
}
