% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastApp.R
\name{contrastApp}
\alias{contrastApp}
\alias{contrastServer}
\alias{contrastInput}
\alias{contrastUI}
\alias{contrastOutput}
\title{Contrast Panel App}
\usage{
contrastApp()

contrastServer(
  id,
  main_par,
  traitSignal,
  traitStats,
  traitModule,
  customSettings = NULL
)

contrastInput(id)

contrastUI(id)

contrastOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Contrast Panel App
}
