% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastGroupApp.R
\name{contrastGroupApp}
\alias{contrastGroupApp}
\alias{contrastGroupServer}
\alias{contrastGroupInput}
\alias{contrastGroupUI}
\alias{contrastGroupOutput}
\title{Groups of Contrasts App}
\usage{
contrastGroupApp()

contrastGroupServer(
  id,
  panel_par,
  main_par,
  traitModule,
  trait_table,
  group_table,
  customSettings = NULL
)

contrastGroupInput(id)

contrastGroupUI(id)

contrastGroupOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitModule}{static data frames}

\item{trait_table, traitContast}{reactive data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Groups of Contrasts App
}
