% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastGroup.R
\name{contrastGroupServer}
\alias{contrastGroupServer}
\alias{contrastGroupInput}
\alias{contrastGroupOutput}
\alias{contrastGroupApp}
\title{Shiny Module Server for Groups of Contrasts}
\usage{
contrastGroupServer(
  id,
  panel_par,
  main_par,
  traitModule,
  groupContrast,
  traitContrast,
  customSettings = NULL
)

contrastGroupInput(id)

contrastGroupOutput(id)

contrastGroupApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitModule}{static data frames}

\item{groupContrast, traitContast}{reactive data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Groups of Contrasts

Shiny Module Input for Groups of Contrasts

Shiny Module Output for Groups of Contrasts

Shiny Module App for Groups of Contrasts
}
