% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastModule.R
\name{contrastModuleServer}
\alias{contrastModuleServer}
\alias{contrastModuleInput}
\alias{contrastModuleOutput}
\alias{contrastModuleApp}
\title{Shiny Module Server for Modules of Contrasts}
\usage{
contrastModuleServer(
  id,
  panel_par,
  main_par,
  traitModule,
  moduleContrast,
  traitContrast,
  customSettings = NULL
)

contrastModuleInput(id)

contrastModuleOutput(id)

contrastModuleApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitModule}{static data frames}

\item{moduleContrast, traitContast}{reactive data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Modules of Contrasts

Shiny Module Input for Modules of Contrasts

Shiny Module Output for Modules of Contrasts

Shiny Module App for Modules of Contrasts
}
