% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPanel.R
\name{contrastPanelServer}
\alias{contrastPanelServer}
\alias{contrastPanelInput}
\alias{contrastPanelUI}
\alias{contrastPanelOutput}
\alias{contrastPanelApp}
\title{Shiny Module Server for Contrast Panel}
\usage{
contrastPanelServer(
  id,
  main_par,
  traitSignal,
  traitStats,
  traitModule,
  customSettings = NULL
)

contrastPanelInput(id)

contrastPanelUI(id)

contrastPanelOutput(id)

contrastPanelApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Contrast Panel

Shiny Module Input for Contrast Panel

Shiny Module Input for Contrast Panel

Shiny Module Output for Contrast Panel

Shiny Module App for Contrast Panel
}
