% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlotApp.R
\name{contrastPlotApp}
\alias{contrastPlotApp}
\alias{contrastPlotServer}
\alias{contrastPlotUI}
\alias{contrastPlotOutput}
\title{Contrast Plots App}
\usage{
contrastPlotApp()

contrastPlotServer(
  id,
  panel_par,
  main_par,
  contrast_table,
  customSettings = NULL,
  modTitle = shiny::reactive("Contrasts")
)

contrastPlotUI(id)

contrastPlotOutput(id)
}
\arguments{
\item{id}{identifier}

\item{panel_par, main_par}{input parameters}

\item{contrast_table}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{modTitle}{character string title for section}
}
\value{
reactive object
}
\description{
Contrast Plots App
}
