% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlot.R
\name{contrastPlotServer}
\alias{contrastPlotServer}
\alias{contrastPlotUI}
\alias{contrastPlotOutput}
\title{Shiny Module Server for Contrast Plots}
\usage{
contrastPlotServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL,
  modTitle = shiny::reactive("Eigentrait Contrasts")
)

contrastPlotUI(id)

contrastPlotOutput(id)
}
\arguments{
\item{id}{identifier}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{modTitle}{character string title for section}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for Contrast Plots

Shiny Module UI for Contrast Plots

Shiny Module Output for Contrast Plots
}
