% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrastServer}
\alias{contrastServer}
\alias{contrastInput}
\alias{contrastUI}
\alias{contrastOutput}
\alias{contrastApp}
\title{Shiny Module Server for Contrast Panel}
\usage{
contrastServer(
  id,
  main_par,
  traitSignal,
  traitStats,
  traitModule,
  customSettings = NULL
)

contrastInput(id)

contrastUI(id)

contrastOutput(id)

contrastApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Contrast Panel

Shiny Module Input for Contrast Panel

Shiny Module Input for Contrast Panel

Shiny Module Output for Contrast Panel

Shiny Module App for Contrast Panel
}
