% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTableApp.R
\name{contrastTableApp}
\alias{contrastTableApp}
\alias{contrastTableServer}
\alias{contrastTableUI}
\title{Contrast Table App}
\usage{
contrastTableApp(id)

contrastTableServer(
  id,
  main_par,
  traitSignal,
  traitStats,
  customSettings = NULL,
  keepDatatraits = shiny::reactive(NULL)
)

contrastTableUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{parameters from calling modules}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}

\item{keepDatatraits}{keep datatraits if not `NULL`}
}
\value{
reactive object
}
\description{
Contrast Table App
}
