% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTable.R
\name{contrastTableServer}
\alias{contrastTableServer}
\alias{contrastTableUI}
\alias{contrastTableApp}
\title{Shiny Module Server for Contrast Panel}
\usage{
contrastTableServer(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitStats,
  customSettings = NULL,
  keepDatatraits = shiny::reactive(NULL)
)

contrastTableUI(id)

contrastTableApp(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{parameters from calling modules}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}

\item{keepDatatraits}{keep datatraits if not `NULL`}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for Contrast Panel

Shiny Module UI for Trait Panel

Shiny Module App for Trait Panel
}
