% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTimeApp.R
\name{contrastTimeApp}
\alias{contrastTimeApp}
\alias{contrastTimeServer}
\alias{contrastTimeInput}
\alias{contrastTimeUI}
\title{Contrasts over Time App}
\usage{
contrastTimeApp()

contrastTimeServer(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitStats,
  contrastTable,
  customSettings = NULL
)

contrastTimeInput(id)

contrastTimeUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Contrasts over Time App

Shiny Module Input for Contrasts over Time

Shiny Module UI for Contrasts over Time
}
