% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTraitApp.R
\name{contrastTraitApp}
\alias{contrastTraitApp}
\alias{contrastTraitServer}
\alias{contrastTraitInput}
\alias{contrastTraitOutput}
\title{Contrast Trait Plots App}
\usage{
contrastTraitApp()

contrastTraitServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL
)

contrastTraitInput(id)

contrastTraitOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Contrast Trait Plots App
}
