% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTrait.R
\name{contrastTraitServer}
\alias{contrastTraitServer}
\alias{contrastTraitUI}
\alias{contrastTraitOutput}
\alias{contrastTraitApp}
\title{Shiny Sex Server for Contrast Plots}
\usage{
contrastTraitServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL
)

contrastTraitUI(id)

contrastTraitOutput(id)

contrastTraitApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Sex Server for Contrast Plots

Shiny Sex UI for Contrast Plots

Shiny Sex Output for Contrast Plots

Shiny Sex App for Contrast Plots
}
