% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{cor_table}
\alias{cor_table}
\title{Correlation Table}
\usage{
cor_table(key_trait, traitSignal, corterm, mincor = 0, reldataset = NULL)
}
\arguments{
\item{key_trait}{}

\item{traitSignal}{}

\item{corterm}{}

\item{mincor}{}

\item{reldataset}{}

\item{data}{frame}
}
\description{
Correlation Table
}
