% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlotApp.R
\name{corPlotApp}
\alias{corPlotApp}
\alias{corPlotServer}
\alias{corPlotOutput}
\title{Correlation Plot App}
\usage{
corPlotApp()

corPlotServer(id, panel_par, cors_table, customSettings = NULL)

corPlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par}{reactive inputs from calling modules}

\item{customSettings}{static list of settings}

\item{input, output, session}{standard shiny arguments}

\item{CorTable}{reactive data frames}
}
\value{
reactive object
}
\description{
Correlation Plot App
}
