% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlotServer}
\alias{corPlotServer}
\alias{corPlotOutput}
\alias{corPlotApp}
\title{Shiny Module Server for Trait Stats}
\usage{
corPlotServer(id, panel_par, main_par, CorTable, customSettings = NULL)

corPlotOutput(id)

corPlotApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive inputs from calling modules}

\item{CorTable}{reactive data frames}

\item{customSettings}{static list of settings}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for Trait Stats

Shiny Module Output for Trait Correlations

Shiny Module App for Trait Correlations
}
