% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTableApp.R
\name{corTableApp}
\alias{corTableApp}
\alias{corTableServer}
\alias{corTableInput}
\alias{corTableOutput}
\title{Correlation Table App}
\usage{
corTableApp()

corTableServer(id, main_par, key_trait, traitSignal, customSettings = NULL)

corTableInput(id)

corTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive inputs from calling modules}

\item{key_trait}{reactive character string}

\item{traitSignal}{static data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Correlation Table App
}
