% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{datasetServer}
\alias{datasetServer}
\alias{datasetInput}
\alias{datasetUI}
\alias{datasetApp}
\title{Shiny Server for dataset INput}
\usage{
datasetServer(id, traitStats = NULL)

datasetInput(id)

datasetUI(id)

datasetApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitStats}{static object}

\item{title}{title of app}
}
\value{
reactive input
}
\description{
Shiny Server for dataset INput
}
