% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotApp.R
\name{dotplotApp}
\alias{dotplotApp}
\alias{dotplotServer}
\alias{dotplotOutput}
\title{DotPlots App}
\usage{
dotplotApp()

dotplotServer(id, panel_par, plot_par, plot_info, contrast_table)

dotplotOutput(id)
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{plot_info}{reactive values from contrastPlot}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object
}
\description{
DotPlots App
}
