% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplotServer}
\alias{dotplotServer}
\alias{dotplotOutput}
\alias{dotplotApp}
\title{Shiny Module Server for DotPlots}
\usage{
dotplotServer(id, panel_par, plot_par, plot_info, contrast_table)

dotplotOutput(id)

dotplotApp()
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{plot_info}{reactive values from contrastPlot}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for DotPlots

Shiny Module Output for Contrast Plots

Shiny App for DotPlot Plots
}
