% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadOutput}
\title{Download App}
\usage{
downloadApp(id)

downloadServer(id, prefix, main_par, download_list)

downloadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{prefix}{static prefix for filename}

\item{main_par}{input parameters from calling routine}

\item{download_list}{reactiveValues with  postfix,plotObject,tableObject}
}
\value{
nothing
}
\description{
Download App
}
