% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadServer}
\alias{downloadServer}
\alias{downloadOutput}
\alias{downloadApp}
\title{Shiny Module Server for Downloads}
\usage{
downloadServer(id, prefix, main_par, download_list)

downloadOutput(id)

downloadApp(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{prefix}{static prefix for filename}

\item{main_par}{input parameters from calling routine}

\item{download_list}{reactiveValues with  postfix,plotObject,tableObject}
}
\value{
nothing

nothing returned

nothing returned
}
\description{
Shiny Module Server for Downloads

Shiny Module Output for Downloads

Shiny Module App for Downloads
}
