% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{eigen_contrast_dataset}
\alias{eigen_contrast_dataset}
\title{Eigen Contrasts from Dataset}
\usage{
eigen_contrast_dataset(object, contr_object)
}
\arguments{
\item{object}{}

\item{contr_object}{}
}
\value{
data frame
}
\description{
Eigen Contrasts from Dataset
}
