% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{eigen_traits_dataset}
\alias{eigen_traits_dataset}
\title{Eigen Traits from Dataset}
\usage{
eigen_traits_dataset(
  object = NULL,
  sexname = NULL,
  modulename = NULL,
  contr_object = NULL,
  eigen_object = eigen_contrast(object, contr_object)
)
}
\arguments{
\item{object}{}

\item{sexname}{}

\item{modulename}{}

\item{contr_object}{}

\item{eigen_object}{}
}
\value{
data frame
}
\description{
Eigen Traits from Dataset
}
