% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entryApp.R
\name{entryApp}
\alias{entryApp}
\alias{entryServer}
\alias{entryInput}
\alias{entryUI}
\alias{entryOutput}
\title{Entry App for password}
\usage{
entryApp()

entryServer(id, customSettings = NULL)

entryInput(id)

entryUI(id)

entryOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{customSettings}{list of custom settings}
}
\value{
reactive server
}
\description{
Entry App for password
}
