% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry.R
\name{entryServer}
\alias{entryServer}
\alias{entryInput}
\alias{entryOutput}
\alias{entryApp}
\title{Shiny Server for entry Key}
\usage{
entryServer(id, customSettings = NULL)

entryInput(id)

entryOutput(id)

entryApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{customSettings}{list of custom settings}

\item{title}{title of app}
}
\value{
reactive server
}
\description{
Shiny Server for entry Key
}
