% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{contrast_signal}
\alias{contrast_signal}
\alias{cor_table}
\alias{eigen_contrast_dataset}
\alias{eigen_traits_dataset}
\alias{is_sex_module}
\alias{mutate_datasets}
\alias{order_choices}
\alias{order_trait_stats}
\alias{select_data_pairs}
\alias{stats_time_table}
\alias{summary_traitTime}
\alias{term_stats}
\alias{time_trait_subset}
\alias{time_units}
\alias{trait_pairs}
\alias{vol_default}
\title{Turn `conditionContrasts` object into a `traitSignal` object.}
\usage{
contrast_signal(contrasts)

cor_table(key_trait, traitSignal, corterm, mincor = 0, reldataset = NULL)

eigen_contrast_dataset(object, contr_object)

eigen_traits_dataset(
  object = NULL,
  sexname = NULL,
  modulename = NULL,
  contr_object = NULL,
  eigen_object = foundr::eigen_contrast(object, contr_object)
)

is_sex_module(object)

mutate_datasets(object, datasets = NULL, undo = FALSE)

order_choices(traitStats)

order_trait_stats(orders, traitStats)

select_data_pairs(object, key_trait, rel_dataset = NULL)

stats_time_table(object, logp = FALSE)

summary_traitTime(object, traitnames = names(object$traits))

term_stats(
  object,
  signal = TRUE,
  condition_name = NULL,
  drop_noise = TRUE,
  cellmean = signal,
  ...
)

time_trait_subset(object, timetrait_all)

time_units(timetrait_all)

trait_pairs(traitnames, sep = " ON ", key = TRUE)

vol_default(ordername)
}
\arguments{
\item{contrasts}{data frame}

\item{key_trait}{name of key trait}

\item{traitSignal}{}

\item{corterm}{}

\item{mincor}{}

\item{reldataset}{}

\item{object}{}

\item{contr_object}{}

\item{sexname}{}

\item{modulename}{}

\item{eigen_object}{}

\item{datasets}{}

\item{undo}{}

\item{traitStats}{data frame with statistics}

\item{orders}{name of order criterion}

\item{rel_dataset}{name of related datasets}

\item{logp}{raise values to power 10 if not `TRUE`}

\item{traitnames}{}

\item{signal}{}

\item{condition_name}{}

\item{drop_noise}{}

\item{cellmean}{}

\item{...}{}

\item{timetrait_all}{data frame}

\item{sep}{}

\item{key}{}

\item{ordername}{name of order column}
}
\value{
data frame

data frame

data frame

data frame

`TRUE` or `FALSE`

data frame with `dataset` and possibly `probandset` columns

vector of stats terms

data frame

data frame

data frame

data frame

vector of terms

data frame

data frame

vector or `trait1 ON trait2`

list of volcano parameters
}
\description{
Turn `conditionContrasts` object into a `traitSignal` object.

Correlation Table

Eigen Contrasts from Dataset

Eigen Traits from Dataset

Is this a sex module?
Apply eigen_contrast over list of `traitModule`s and `contr_object`.

Mutate Datasets

Order Choices

Order Trait Statistics

Select data including Key Trait and Related Datasets

Stats Time Table

Summary fo traitTime Object

Stat Terms

Time Trait Subset

Time Units

Trait Pairs

Volcano Defaults
}
