% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundrApp.R
\name{foundrApp}
\alias{foundrApp}
\alias{foundrServer}
\alias{foundrInput}
\alias{foundrOutput}
\title{foundr App for foundr package}
\usage{
foundrApp()

foundrServer(
  id,
  traitData = NULL,
  traitSignal = NULL,
  traitStats = NULL,
  customSettings = NULL,
  traitModule = NULL
)

foundrInput(id)

foundrOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}
}
\value{
reactive server
}
\description{
foundr App for foundr package
}
