% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr.R
\name{foundrServer}
\alias{foundrServer}
\alias{foundrInput}
\alias{foundrOutput}
\alias{foundrApp}
\title{Shiny Server for foundr Package}
\usage{
foundrServer(
  id,
  traitData = NULL,
  traitSignal = NULL,
  traitStats = NULL,
  customSettings = NULL,
  traitModule = NULL
)

foundrInput(id)

foundrOutput(id)

foundrApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}

\item{title}{title of app}
}
\value{
reactive server
}
\description{
Shiny Server for foundr Package
}
