% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundrSetup.R
\name{foundrSetup}
\alias{foundrSetup}
\title{Setup Data for Foundr App}
\usage{
foundrSetup(instance = c("Liver", "Trait"), custom = FALSE)
}
\arguments{
\item{instance}{type of data}

\item{custom}{setup custom if `TRUE`}
}
\value{
invisible
}
\description{
Setup Data for Foundr App
}
