% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainParApp.R
\name{mainParApp}
\alias{mainParApp}
\alias{mainParServer}
\alias{mainParInput}
\alias{mainParUI}
\alias{mainParOutput}
\title{mainPar App}
\usage{
mainParApp()

mainParServer(id, traitStats = NULL)

mainParInput(id)

mainParUI(id)

mainParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitStats}{static object}
}
\value{
reactive input
}
\description{
mainPar App
}
