% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainPar.R
\name{mainParServer}
\alias{mainParServer}
\alias{mainParInput}
\alias{mainParUI}
\alias{mainParApp}
\title{Shiny Server for mainPar Server}
\usage{
mainParServer(id, traitStats = NULL)

mainParInput(id)

mainParUI(id)

mainParApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitStats}{static object}

\item{title}{title of app}
}
\value{
reactive input
}
\description{
Shiny Server for mainPar Server
}
