% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{mutate_datasets}
\alias{mutate_datasets}
\title{Mutate Datasets}
\usage{
mutate_datasets(object, datasets = NULL, undo = FALSE)
}
\arguments{
\item{object}{}

\item{datasets}{}

\item{undo}{}
}
\value{
data frame with `dataset` and possibly `probandset` columns
}
\description{
Mutate Datasets
}
