% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{order_trait_stats}
\alias{order_trait_stats}
\title{Order Trait Statistics}
\usage{
order_trait_stats(orders, traitStats)
}
\arguments{
\item{orders}{name of order criterion}

\item{traitStats}{data frame with statistics}
}
\value{
data frame
}
\description{
Order Trait Statistics
}
