% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelParApp.R
\name{panelParApp}
\alias{panelParApp}
\alias{panelParServer}
\alias{panelParInput}
\alias{panelParUI}
\alias{panelParOutput}
\title{Panel Parameters App}
\usage{
panelParApp()

panelParServer(id, main_par, traitStats = NULL, panel_name = NULL)

panelParInput(id)

panelParUI(id)

panelParOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{main parameters}

\item{traitStats}{static object}

\item{panel_name}{name of panel}
}
\value{
reactive input
}
\description{
Panel Parameters App
}
