% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPar.R
\name{panelParServer}
\alias{panelParServer}
\alias{panelParStrains}
\alias{panelParTraits}
\alias{panelParSex}
\alias{panelParApp}
\title{Shiny Server for panelPar Server}
\usage{
panelParServer(id, main_par, traitSignal = NULL)

panelParStrains(id)

panelParTraits(id)

panelParSex(id)

panelParApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{title}{title of app}

\item{traitStats}{static object}
}
\value{
reactive input
}
\description{
Shiny Server for panelPar Server
}
