% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParApp.R
\name{plotParApp}
\alias{plotParApp}
\alias{plotParServer}
\alias{plotParInput}
\alias{plotParUI}
\alias{plotParOutput}
\title{Plot Parameters App}
\usage{
plotParApp()

plotParServer(id, contrast_table)

plotParInput(id)

plotParUI(id)

plotParOutput(id)
}
\arguments{
\item{id}{identifier}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object
}
\description{
Plot Parameters App
}
