% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPar.R
\name{plotParServer}
\alias{plotParServer}
\alias{plotParInput}
\alias{plotParUI}
\alias{plotParApp}
\title{Shiny Module Server for Plot Parameters}
\usage{
plotParServer(id, contrastTable)

plotParInput(id)

plotParUI(id)

plotParApp()
}
\arguments{
\item{id}{identifier}

\item{contrastTable}{reactive data frame}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Plot Parameters

Shiny Module Input for Plot Parameters

Shiny Module UI for Plot Parameters

Shiny Module App for Plot Parameters
}
