% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{stats_time_table}
\alias{stats_time_table}
\title{Stats Time Table}
\usage{
stats_time_table(object, logp = FALSE)
}
\arguments{
\item{object}{list of data frames}

\item{logp}{raise values to power 10 if not `TRUE`}
}
\value{
data frame
}
\description{
Stats Time Table
}
