% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsApp.R
\name{statsApp}
\alias{statsApp}
\alias{statsServer}
\alias{statsUI}
\alias{statsOutput}
\title{Stats App}
\usage{
statsApp()

statsServer(id, main_par, traitStats, customSettings = NULL, facet = FALSE)

statsUI(id)

statsOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments from `server()`}

\item{traitStats}{static data frame}

\item{customSettings}{list of custom settings}

\item{facet}{facet on `strain` if `TRUE`}
}
\value{
reactive object for `statsOutput`

nothing returned
}
\description{
Stats App

Shiny Module Output for Stats Plot
}
