% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsPanel.R
\name{statsPanelServer}
\alias{statsPanelServer}
\alias{statsPanelOutput}
\alias{statsPanelApp}
\title{Shiny Module Server for Stats Plots}
\usage{
statsPanelServer(
  id,
  main_par,
  traitStats,
  customSettings = NULL,
  facet = FALSE
)

statsPanelOutput(id)

statsPanelApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments from `server()`}

\item{traitStats}{static data frame}

\item{customSettings}{list of custom settings}

\item{facet}{facet on `strain` if `TRUE`}
}
\value{
reactive object for `statsPanelOutput`

nothing returned

nothing returned
}
\description{
Shiny Module Server for Stats Plots

Shiny Module Output for Stats Plot

Shiny Module App for Stats Plot
}
