% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statsServer}
\alias{statsServer}
\alias{statsOutput}
\alias{statsApp}
\title{Shiny Module Server for Stats Plots}
\usage{
statsServer(id, main_par, traitStats, customSettings = NULL, facet = FALSE)

statsOutput(id)

statsApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments from `server()`}

\item{traitStats}{static data frame}

\item{customSettings}{list of custom settings}

\item{facet}{facet on `strain` if `TRUE`}
}
\value{
reactive object for `statsOutput`

nothing returned

nothing returned
}
\description{
Shiny Module Server for Stats Plots

Shiny Module Output for Stats Plot

Shiny Module App for Stats Plot
}
