% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{summary_traitTime}
\alias{summary_traitTime}
\title{Summary fo traitTime Object}
\usage{
summary_traitTime(object, traitnames = names(object$traits))
}
\arguments{
\item{object}{data frame}

\item{traitnames}{character list}
}
\value{
data frame
}
\description{
Summary fo traitTime Object
}
