% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabServer}
\alias{tabServer}
\alias{tabInput}
\alias{tabOutput}
\alias{tabApp}
\title{Shiny tab Server for foundr Package}
\usage{
tabServer(
  id,
  traitData = NULL,
  traitSignal = NULL,
  traitStats = NULL,
  customSettings = NULL,
  traitModule = NULL
)

tabInput(id)

tabOutput(id)

tabApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}

\item{title}{title of app}
}
\value{
reactive server
}
\description{
Shiny tab Server for foundr Package
}
