% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{term_stats}
\alias{term_stats}
\title{Stat Terms}
\usage{
term_stats(
  object,
  signal = TRUE,
  condition_name = NULL,
  drop_noise = TRUE,
  cellmean = signal,
  ...
)
}
\arguments{
\item{object}{}

\item{signal}{}

\item{condition_name}{}

\item{drop_noise}{}

\item{cellmean}{}

\item{...}{}
}
\value{
vector of terms
}
\description{
Stat Terms
}
