% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{time_trait_subset}
\alias{time_trait_subset}
\title{Time Trait Subset}
\usage{
time_trait_subset(object, timetrait_all)
}
\arguments{
\item{object}{}

\item{timetrait_all}{}
}
\value{
data frame
}
\description{
Time Trait Subset
}
