% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeApp.R
\name{timeApp}
\alias{timeApp}
\alias{timeServer}
\alias{timeInput}
\alias{timeUI}
\alias{timeOutput}
\title{Times Panel App}
\usage{
timeApp()

timeServer(id, main_par, traitData, traitSignal, traitStats)

timeInput(id)

timeUI(id)

timeOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned
}
\description{
Times Panel App
}
