% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePanel.R
\name{timePanelServer}
\alias{timePanelServer}
\alias{timePanelInput}
\alias{timePanelUI}
\alias{timePanelOutput}
\alias{timePanelApp}
\title{Shiny Module Server for Times Plots}
\usage{
timePanelServer(id, main_par, traitData, traitSignal, traitStats)

timePanelInput(id)

timePanelUI(id)

timePanelOutput(id)

timePanelApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Times Plots

Shiny Module Input for Time Panel

Shiny Module UI for Time Panel

Shiny Module Output for Times Plot

Shiny Module App for Times Plot
}
