% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlotApp.R
\name{timePlotApp}
\alias{timePlotApp}
\alias{timePlotServer}
\alias{timePlotOutput}
\title{Time Plots App}
\usage{
timePlotApp()

timePlotServer(id, panel_par, main_par, traitSignal, time_table)

timePlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{time_table}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned
}
\description{
Time Plots App
}
