% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlot.R
\name{timePlotServer}
\alias{timePlotServer}
\alias{timePlotInput}
\alias{timePlotUI}
\alias{timePlotOutput}
\title{Shiny Module Server for Time Plots}
\usage{
timePlotServer(id, panel_par, main_par, traitSignal, traitTimesData)

timePlotInput(id)

timePlotUI(id)

timePlotOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitTimesData}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Time Plots

Shiny Module UI for Time Plots

Shiny Module UI for Time Plots

Shiny Module Output for Time Plots
}
