% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeServer}
\alias{timeServer}
\alias{timeInput}
\alias{timeUI}
\alias{timeOutput}
\alias{timeApp}
\title{Shiny Module Server for Times Plots}
\usage{
timeServer(id, main_par, traitData, traitSignal, traitStats)

timeInput(id)

timeUI(id)

timeOutput(id)

timeApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Times Plots

Shiny Module Input for Time Panel

Shiny Module UI for Time Panel

Shiny Module Output for Times Plot

Shiny Module App for Times Plot
}
