% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTableApp.R
\name{timeTableApp}
\alias{timeTableApp}
\alias{timeTableServer}
\alias{timeTableInput}
\alias{timeTableUI}
\alias{timeTableOutput}
\title{Times Tables App}
\usage{
timeTableApp()

timeTableServer(id, panel_par, main_par, traitData, traitSignal, traitStats)

timeTableInput(id)

timeTableUI(id)

timeTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned
}
\description{
Times Tables App
}
