% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{timeTableServer}
\alias{timeTableServer}
\alias{timeTableInput}
\alias{timeTableUI}
\alias{timeTableOutput}
\title{Shiny Module Server for Times Plots}
\usage{
timeTableServer(id, panel_par, main_par, traitData, traitSignal, traitStats)

timeTableInput(id)

timeTableUI(id)

timeTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Times Plots

Shiny Module Input for Times Table

Shiny Module UI for Times Table

Shiny Module Output for Times Table
}
