% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTraitsApp.R
\name{timeTraitsApp}
\alias{timeTraitsApp}
\alias{timeTraitsServer}
\alias{timeTraitsInput}
\alias{timeTraitsUI}
\alias{timeTraitsOutput}
\title{Times Traits App}
\usage{
timeTraitsApp()

timeTraitsServer(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitOrder,
  responses = c("value", "normed", "cellmean")
)

timeTraitsInput(id)

timeTraitsUI(id)

timeTraitsOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitOrder}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned
}
\description{
Times Traits App
}
