% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTraits.R
\name{timeTraitsServer}
\alias{timeTraitsServer}
\alias{timeTraitsInput}
\alias{timeTraitsUI}
\alias{timeTraitsOutput}
\alias{timeTraitsApp}
\title{Shiny Module Server for Times Plots}
\usage{
timeTraitsServer(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitOrder,
  responses = c("value", "normed", "cellmean")
)

timeTraitsInput(id)

timeTraitsUI(id)

timeTraitsOutput(id)

timeTraitsApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitOrder}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Times Plots

Shiny Module Input for Time Traits

Shiny Module UI for Time Traits

Shiny Module Output for Time Traits

Shiny Module App for Times Plot
}
