% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{trait_pairs}
\alias{trait_pairs}
\title{Trait Pairs}
\usage{
trait_pairs(traitnames, sep = " ON ", key = TRUE)
}
\arguments{
\item{traitnames}{}

\item{sep}{}

\item{key}{}
}
\value{
vector or `trait1 ON trait2`
}
\description{
Trait Pairs
}
