% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitApp.R
\name{traitApp}
\alias{traitApp}
\alias{traitServer}
\alias{traitInput}
\alias{traitUI}
\alias{traitOutput}
\title{Trait Panel App}
\usage{
traitApp()

traitServer(
  id,
  main_par,
  traitData,
  traitSignal,
  traitStats,
  customSettings = NULL
)

traitInput(id)

traitUI(id)

traitOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Trait Panel App
}
