% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitNamesApp.R
\name{traitNamesApp}
\alias{traitNamesApp}
\alias{traitNamesServer}
\alias{traitNamesUI}
\title{Trait Names App}
\usage{
traitNamesApp()

traitNamesServer(id, main_par, traitArranged, multiples = FALSE)

traitNamesUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitArranged}{reactive data frames}

\item{multiples}{fixed logical for multiple trait names}
}
\value{
reactive vector of trait names
}
\description{
Select trait names in one of two modes, depending on the fixed `multiples`:
`FALSE` = only one trait name,
`TRUE` =  multiple names.
The order of choices depends on `traitArranged()`.
}
