% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitNames.R
\name{traitNamesServer}
\alias{traitNamesServer}
\alias{traitNamesUI}
\alias{traitNamesApp}
\title{Shiny Module Server for Trait Names}
\usage{
traitNamesServer(id, main_par, traitArranged, multiples = FALSE)

traitNamesUI(id)

traitNamesApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitArranged}{reactive data frames}

\item{multiples}{fixed logical for multiple trait names}
}
\value{
reactive vector of trait names

nothing returned

nothing returned
}
\description{
Select trait names in one of two modes, depending on the fixed `multiples`:
`FALSE` = only one trait name,
`TRUE` =  multiple names.
The order of choices depends on `traitArranged()`.
}
