% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitOrderApp.R
\name{traitOrderApp}
\alias{traitOrderApp}
\alias{traitOrderServer}
\alias{traitOrderUI}
\title{Trait Order App}
\usage{
traitOrderApp()

traitOrderServer(
  id,
  main_par,
  traitStats,
  customSettings = NULL,
  keepDatatraits = shiny::reactive(NULL)
)

traitOrderUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{input reactive list}

\item{traitStats}{static data frame}

\item{customSettings}{custom settings list}

\item{keepDatatraits}{keep datatraits if not `NULL`}
}
\value{
reactive object
}
\description{
Trait Order App
}
