% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitPairsApp.R
\name{traitPairsApp}
\alias{traitPairsApp}
\alias{traitPairsServer}
\alias{traitPairsOutput}
\title{Trait Pairs App}
\usage{
traitPairsApp(id)

traitPairsServer(id, panel_par, main_par, trait_names, trait_table)

traitPairsOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments from `foundrServer`}

\item{trait_names}{reactive with trait names}

\item{trait_table}{reactive objects from `foundrServer`}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object for `traitSolos`
}
\description{
Trait Pairs App
}
