% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitPairs.R
\name{traitPairsServer}
\alias{traitPairsServer}
\alias{traitPairsUI}
\alias{traitPairsApp}
\title{Shiny Module Server for trait solos Plots}
\usage{
traitPairsServer(id, panel_par, main_par, trait_names, traitSolosObject)

traitPairsUI(id)

traitPairsApp(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments from `foundrServer`}

\item{trait_names}{reactive with trait names}

\item{traitSolosObject}{reactive objects from `foundrServer`}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object for `traitSolos`

nothing returned

nothing returned
}
\description{
Shiny Module Server for trait solos Plots

Shiny Module UI for TraitPairs

Shiny Module App for TraitPairs
}
