% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitPanel.R
\name{traitPanelServer}
\alias{traitPanelServer}
\alias{traitPanelInput}
\alias{traitPanelUI}
\alias{traitPanelOutput}
\alias{traitPanelApp}
\title{Shiny Module Server for Trait Panel}
\usage{
traitPanelServer(
  id,
  main_par,
  traitData,
  traitSignal,
  traitStats,
  customSettings = NULL
)

traitPanelInput(id)

traitPanelUI(id)

traitPanelOutput(id)

traitPanelApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Trait Panel

Shiny Module Input for Trait Panel

Shiny Module UI for Trait Panel

Shiny Module Output for Trait Panel

Shiny Module App for Trait Panel
}
