% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitSolosApp.R
\name{traitSolosApp}
\alias{traitSolosApp}
\alias{traitSolosServer}
\alias{traitSolosOutput}
\title{Trait Solos App}
\usage{
traitSolosApp()

traitSolosServer(id, panel_par, main_par, trait_table)

traitSolosOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments from `server`}

\item{trait_table}{reactive objects from `server`}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object
}
\description{
Trait Solos App
}
