% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitSolos.R
\name{traitSolosServer}
\alias{traitSolosServer}
\alias{traitSolosOutput}
\alias{traitSolosApp}
\title{Shiny Module Server for trait solos Plots}
\usage{
traitSolosServer(id, panel_par, main_par, trait_table)

traitSolosOutput(id)

traitSolosApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments from `server`}

\item{trait_table}{reactive objects from `server`}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for trait solos Plots

Shiny Module Output for traitSolos

Shiny Module AI for traitSolos
}
