% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitTableApp.R
\name{traitTableApp}
\alias{traitTableApp}
\alias{traitTableServer}
\alias{traitTableUI}
\alias{traitTableOutput}
\title{Trait Table App}
\usage{
traitTableApp()

traitTableServer(
  id,
  panel_par,
  key_trait,
  rel_traits,
  traitData,
  traitSignal,
  customSettings = NULL
)

traitTableUI(id)

traitTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par}{reactive arguments}

\item{key_trait, rel_traits}{reactives with trait names}

\item{traitData, traitSignal}{static objects}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object
}
\description{
Trait Table App
}
