% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitTable.R
\name{traitTableServer}
\alias{traitTableServer}
\alias{traitTableUI}
\alias{traitTableOutput}
\alias{traitTableApp}
\title{Shiny Module Server for Trait Table}
\usage{
traitTableServer(
  id,
  panel_par,
  key_trait,
  rel_traits,
  traitData,
  traitSignal,
  customSettings = NULL
)

traitTableUI(id)

traitTableOutput(id)

traitTableApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par}{reactive arguments}

\item{key_trait, rel_traits}{reactives with trait names}

\item{traitData, traitSignal}{static objects}

\item{input, output, session}{standard shiny arguments}
}
\value{
reactive object for `shinyTrait` routines

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Trait Table

Shiny Module UI for Trait Table

Shiny Module UI for Trait Table

Shiny Module App for Trait Table
}
