% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundr_helpers.R
\name{vol_default}
\alias{vol_default}
\title{Volcano Defaults}
\usage{
vol_default(ordername)
}
\arguments{
\item{ordername}{name of order column}
}
\value{
list of volcano parameters
}
\description{
Volcano Defaults
}
