% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoApp.R
\name{volcanoApp}
\alias{volcanoApp}
\alias{volcanoServer}
\alias{volcanoOutput}
\title{Volcano Plot App}
\usage{
volcanoApp()

volcanoServer(id, panel_par, plot_par, plot_info, contrast_table)

volcanoOutput(id)
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{plot_info}{reactive values from contrastPlot}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object
}
\description{
Volcano Plot App
}
