% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoServer}
\alias{volcanoServer}
\alias{volcanoInput}
\alias{volcanoOutput}
\alias{volcanoApp}
\title{Shiny Module Server for Contrast Plots}
\usage{
volcanoServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL,
  modTitle = shiny::reactive("Eigentrait Contrasts")
)

volcanoInput(id)

volcanoOutput(id)

volcanoApp()
}
\arguments{
\item{id}{identifier}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{modTitle}{character string title for section}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Contrast Plots

Shiny Module Input for Volcano Plots

Shiny Module Output for Contrast Plots

Shiny App for Volcano Plots
}
