% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplotServer}
\alias{biplotServer}
\alias{biplotOutput}
\alias{biplotApp}
\title{Shiny Module Server for BiPlots}
\usage{
biplotServer(
  id,
  panel_par,
  plot_par,
  contrast_table,
  info,
  filter_rownames,
  threshold
)

biplotOutput(id)

biplotApp()
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{contrast_table}{reactive data frame}

\item{info, filter_rownames, threshold}{reactive items from contrastPlot}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for BiPlots

Shiny Module Output for Contrast Plots

Shiny App for BiPlot Plots
}
