% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastSex.R
\name{contrastSexServer}
\alias{contrastSexServer}
\alias{contrastSexUI}
\alias{contrastSexOutput}
\alias{contrastSexApp}
\title{Shiny Sex Server for Contrast Plots}
\usage{
contrastSexServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL
)

contrastSexUI(id)

contrastSexOutput(id)

contrastSexApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Sex Server for Contrast Plots

Shiny Sex UI for Contrast Plots

Shiny Sex Output for Contrast Plots

Shiny Sex App for Contrast Plots
}
