% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastTrait.R
\name{contrastTraitServer}
\alias{contrastTraitServer}
\alias{contrastTraitInput}
\alias{contrastTraitOutput}
\alias{contrastTraitApp}
\title{Shiny Trait Server for Contrast Plots}
\usage{
contrastTraitServer(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL
)

contrastTraitInput(id)

contrastTraitOutput(id)

contrastTraitApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Trait Server for Contrast Plots

Shiny Trait Input for Contrast Plots

Shiny Trait Output for Contrast Plots

Shiny Trait App for Contrast Plots
}
