% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTable.R
\name{corTableServer}
\alias{corTableServer}
\alias{corTableInput}
\alias{corTableOutput}
\alias{corTableApp}
\title{Shiny Module Server for Correlation Table}
\usage{
corTableServer(id, main_par, key_trait, traitSignal, customSettings = NULL)

corTableInput(id)

corTableOutput(id)

corTableApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive inputs from calling modules}

\item{key_trait}{reactive character string}

\item{traitSignal}{static data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive object

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Server for Correlation Table

Shiny Module UI for Trait Correlations

Shiny Module Output for Trait Correlations

Shiny Module App for Trait Correlations
}
