% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadServer}
\alias{downloadServer}
\alias{downloadOutput}
\title{Shiny Module Server for Downloads}
\usage{
downloadServer(id, prefix, main_par, postfix, plotObject, tableObject)

downloadOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{prefix}{static prefix for filename}

\item{main_par}{input parameters from calling routine}

\item{postfix}{reactive postfix for filename}

\item{tableObject}{reactive table pring object}

\item{plotObjec}{reactive plot print object}
}
\value{
nothing

nothing returned
}
\description{
Shiny Module Server for Downloads

Shiny Module Output for Downloads
}
