% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPar.R
\name{panelParServer}
\alias{panelParServer}
\alias{panelParInput}
\alias{panelParUI}
\alias{panelParApp}
\title{Shiny Server for panelPar Server}
\usage{
panelParServer(id, main_par, traitStats = NULL)

panelParInput(id)

panelParUI(id)

panelParApp(title = "")
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitStats}{static object}

\item{title}{title of app}
}
\value{
reactive input
}
\description{
Shiny Server for panelPar Server
}
