% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTable.R
\name{timeTableServer}
\alias{timeTableServer}
\alias{timeTableInput}
\alias{timeTableUI}
\alias{timeTableOutput}
\alias{timeTableApp}
\title{Shiny Server for Times Plots}
\usage{
timeTableServer(id, panel_par, main_par, traitData, traitSignal, traitStats)

timeTableInput(id)

timeTableUI(id)

timeTableOutput(id)

timeTableApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Times Plots

Shiny Input for Times Table

Shiny UI for Times Table

Shiny Output for Times Table

Shiny App for Times Table
}
