% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitOrder.R
\name{traitOrderServer}
\alias{traitOrderServer}
\alias{traitOrderApp}
\title{Shiny Module Server for Trait Order}
\usage{
traitOrderServer(
  id,
  main_par,
  traitStats,
  customSettings = NULL,
  keepDatatraits = shiny::reactive(NULL)
)

traitOrderApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{input reactive list}

\item{traitStats}{static data frame}

\item{customSettings}{custom settings list}

\item{keepDatatraits}{keep datatraits if not `NULL`}
}
\value{
reactive object
}
\description{
Shiny Module Server for Trait Order

Shiny Module App for Trait Order
}
