% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoServer}
\alias{volcanoServer}
\alias{volcanoOutput}
\alias{volcanoApp}
\title{Shiny Module Server for Volcano Plots}
\usage{
volcanoServer(
  id,
  panel_par,
  plot_par,
  contrast_table,
  info,
  filter_rownames,
  threshold
)

volcanoOutput(id)

volcanoApp()
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{contrast_table}{reactive data frame}

\item{info, filter_rownames, threshold}{reactive items from contrastPlot}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for Volcano Plots

Shiny Module Output for Volcano Plots

Shiny App for Volcano Plots
}
