% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoServer}
\alias{volcanoServer}
\alias{volcanoOutput}
\alias{volcanoApp}
\title{Shiny Module Server for Volcano Plots}
\usage{
volcanoServer(id, panel_par, plot_par, plot_info, contrast_table)

volcanoOutput(id)

volcanoApp()
}
\arguments{
\item{id}{identifier}

\item{panel_par, plot_par}{input parameters}

\item{plot_info}{reactive values from contrastPlot}

\item{contrast_table}{reactive data frame}
}
\value{
reactive object

nothing returned

nothing returned
}
\description{
Shiny Module Server for Volcano Plots

Shiny Module Output for Volcano Plots

Shiny App for Volcano Plots
}
